begin better organization of the repo Separate all the interfaces
