+++ sharder.py
@@ -1,4 +1,5 @@
 import math
+import os
 
 # global SHARD_MULTIPLES_BACK, MAX_SHARD_SIZE
